+++ dtslint/ts4.1/ReaderT.ts
@@ -33,19 +33,19 @@ _.chain(RTE.Chain)
 //
 
 // $ExpectType <R, S, FR, W, E, A>(f: (r: R) => Either<E, A>) => Reader<R, TaskEither<E, A>>
-_.fromNaturalTransformation<E.EitherF, TE.TaskEitherF>(TE.fromEither)
+_.fromFunctionK<E.EitherF, TE.TaskEitherF>(TE.fromEither)
 
 // $ExpectType <R, S, FR, W, E, A>(f: (r: R) => IO<A>) => Reader<R, TaskEither<E, A>>
-_.fromNaturalTransformation<IO.IOF, TE.TaskEitherF>(TE.fromIO)
+_.fromFunctionK<IO.IOF, TE.TaskEitherF>(TE.fromIO)
 
 // $ExpectType <R, S, FR, W, E, A>(f: (r: R) => Reader<FR, A>) => Reader<R, ReaderTaskEither<FR, E, A>>
-_.fromNaturalTransformation<R.ReaderF, RTE.ReaderTaskEitherF>(RTE.fromReader)
+_.fromFunctionK<R.ReaderF, RTE.ReaderTaskEitherF>(RTE.fromReader)
 
 // $ExpectType <R, S, FR, W, E, A>(f: (r: R) => State<S, A>) => Reader<R, StateReaderTaskEither<S, FR, E, A>>
-_.fromNaturalTransformation<S.StateF, SRTE.StateReaderTaskEitherF>(SRTE.fromState)
+_.fromFunctionK<S.StateF, SRTE.StateReaderTaskEitherF>(SRTE.fromState)
 
 // $ExpectType <R, S, FR, W, E, A>(f: (r: R) => Task<A>) => Reader<R, TaskEither<E, A>>
-_.fromNaturalTransformation<T.TaskF, TE.TaskEitherF>(TE.fromTask)
+_.fromFunctionK<T.TaskF, TE.TaskEitherF>(TE.fromTask)
 
 // $ExpectType <R, S, FR, W, E, A>(f: (r: R) => These<E, A>) => Reader<R, TaskThese<E, A>>
-_.fromNaturalTransformation<TH.TheseF, TTH.TaskTheseF>(TTH.fromThese)
+_.fromFunctionK<TH.TheseF, TTH.TaskTheseF>(TTH.fromThese)