+++ src/FunctionK.ts
@@ -14,6 +14,6 @@ import type { HKT, Kind } from './HKT'
 /**
  * @since 3.0.0
  */
-export interface NaturalTransformation<F extends HKT, G extends HKT> {
+export interface FunctionK<F extends HKT, G extends HKT> {
   <S, R, W, E, A>(fa: Kind<F, S, R, W, E, A>): Kind<G, S, R, W, E, A>
 }