+++ test/ReaderT.ts
@@ -5,8 +5,8 @@ import * as TE from '../src/TaskEither'
 import * as U from './util'
 
 describe('ReaderT', () => {
-  it('fromNaturalTransformation', async () => {
-    const fromReaderIO = _.fromNaturalTransformation<IO.IOF, TE.TaskEitherF>(TE.fromIO)
+  it('fromFunctionK', async () => {
+    const fromReaderIO = _.fromFunctionK<IO.IOF, TE.TaskEitherF>(TE.fromIO)
     const f = (s: string): IO.IO<number> => IO.of(s.length)
     const fa = fromReaderIO(f)
     U.deepStrictEqual(await fa('a')(), E.right(1))